Add module hierarchy with shorter module names
Addresses #110.
