+++ prettyprinter/src/Data/Text/Prettyprint/Doc/Internal/Type.hs
@@ -1,14 +1,12 @@
-{-# OPTIONS_HADDOCK not-home #-}
-
--- | __Internal module with stability guarantees__
+-- | This module is part of the old @prettyprinter@ API which is planned to be
+-- deprecated and eventually removed.
 --
--- This module exposes the internals of the @'Doc'@ type so other libraries can
--- write adaptors to/from it. For all other uses, please use only the API
--- provided by non-internal modules.
---
--- Although this module is internal, it follows the usual package versioning
--- policy, AKA Haskell’s version of semantic versioning. In other words, this
--- module is as stable as the public API.
-module Data.Text.Prettyprint.Doc.Internal.Type (Doc(..)) where
+-- Use "Prettyprinter.Internal.Type" instead.
+module Data.Text.Prettyprint.Doc.Internal.Type (
+    module Prettyprinter.Internal.Type
+) where
 
-import Data.Text.Prettyprint.Doc.Internal
+import Prettyprinter.Internal.Type
+
+
+
