+++ prettyprinter/src/Prettyprinter/Render/String.hs
@@ -1,0 +1,10 @@
+module Prettyprinter.Render.String (
+    renderString,
+    renderShowS,
+) where
+
+import Prettyprinter.Internal (SimpleDocStream, renderShowS)
+
+-- | Render a 'SimpleDocStream' to a 'String'.
+renderString :: SimpleDocStream ann -> String
+renderString s = renderShowS s ""
